+++ src/lib/constants.ts
@@ -5,3 +5,5 @@ export const PROTONS_ATOMS_REQUIRED = 1_000_000_000;
 export const ELECTRONS_PROTONS_REQUIRED = 1_000_000_000;
 
 export const XP_PER_ATOM = 0.1;
+
+export const MAX_BOOST_POINTS = 20;